// public/types.ts
/*
 * Licensed to Elasticsearch B.V. under one or more contributor
 * license agreements. See the NOTICE file distributed with
 * this work for additional information regarding copyright
 * ownership. Elasticsearch B.V. licenses this file to you under
 * the Apache License, Version 2.0 (the "License"); you may
 * not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

import { TimefilterContract, FilterManager } from '../../../src/plugins/data/public';
import { CoreSetup } from '../../../src/core/public';

export interface CustomFormFilterAccountsVisParams {
  cvs_claim_nbr: string;
  business_ln: string;
  file_id: string;
  ndc_cd: string;
  drug_label: string;
  rx_member_id: string;
  rptng_mbr_id: string;
  nabp_nbr: string;
  prescriber_npi: string;
  group_nbr: string;
  filterCounter: number;
  timefilter: TimefilterContract;
  filterManager: FilterManager;
  coreSetup: CoreSetup;
}